migrate integration testing to new crate
