+++ actix-test/CHANGES.md
@@ -1,0 +1,6 @@
+# Changes
+
+## Unreleased - 2021-xx-xx
+* Move integration testing structs from `actix-web`. [#???]
+
+[#???]: https://github.com/actix/actix-web/pull/???
