+++ codecov.yml
@@ -11,6 +11,6 @@ coverage:
 
 ignore: # ignore code coverage on following paths
   - "**/tests"
-  - "test-server"
   - "**/benches"
   - "**/examples"
+  - "test-server"
